requirements.txt
unbabel-comet==2.2.1
sacrebleu==2.4.3
scikit-learn==1.5.1
tqdm==4.66.5
